% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_plot_g}
\alias{create_plot_g}
\title{Creates a plot for the input model}
\usage{
create_plot_g(mod, input_frame)
}
\arguments{
\item{mod}{a model object}

\item{input_frame}{a data frame containing input data for the model}
}
\value{
A plot of predicted values for avacc and ig parameters for males and females
}
\description{
The function takes a model object and an input frame to create a plot of predicted values.
The plot shows the predicted values for two parameters (avacc and ig) for males and females.
It also shows the percentiles of the predicted values for each parameter and sex
}
\examples{
mod <- lm(avacc_m ~ age, data = input_data)
input_frame <- data.frame(parameter = c("avacc", "avacc", "ig", "ig"), gender = c("m", "f", "m", "f"),
age = c(20, 25, 30, 35), values = c(2.5, 3.1, 1.5, 2.0))
create_plot_g(mod, input_frame)
}
