% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_plot_i}
\alias{create_plot_i}
\title{Create a publication-quality plot}
\usage{
create_plot_i(mod, input_frame)
}
\arguments{
\item{mod}{A model object that includes the predictions for the average acceleration and the intensity gradient.}

\item{input_frame}{A data frame that includes the measured values for the average acceleration and the intensity gradient, as well as the age of the participants.}
}
\value{
The function returns a publication-quality plot.
}
\description{
This function creates a publication-quality plot with two subplots. The first subplot shows the average acceleration in mg over different percentiles of age, and the second subplot shows the intensity gradient over different percentiles of age. The function takes two arguments:
}
\details{
\code{mod}: A model object that includes the predictions for the average acceleration and the intensity gradient. The object should have two model components with the strings "avacc" and "ig" in their names, respectively. The function uses \code{create_predframe} to create the data frames for each of the subplots.

\code{input_frame}: A data frame that includes the measured values for the average acceleration and the intensity gradient, as well as the age of the participants. The data frame should have three columns: "age", "parameter", and "value".
}
\examples{
create_plot_i(mod = my_model, input_frame = my_data_frame)
}
