% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_predframe}
\alias{create_predframe}
\title{Predict centiles and plot them using ggplot2}
\usage{
create_predframe(mod)
}
\arguments{
\item{mod}{a statistical model object.}
}
\value{
a data.frame containing the predicted centiles.
}
\description{
This function generates the centiles predicted by a model and plots them using ggplot2.
}
\examples{
mod <- lm(mpg ~ wt + cyl + hp, data = mtcars)
create_predframe(mod)
}
