% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R, R/www.R
\name{.onLoad}
\alias{.onLoad}
\title{This function adds a resource path for the shiny application.}
\usage{
.onLoad(libname, pkgname)

.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the package's library where the resource path will be added.}

\item{pkgname}{The name of the package where the resource path will be added.}
}
\description{
This function adds a resource path for the shiny application.

This function adds a resource path for the shiny application.
}
\details{
This function adds a resource path for the shiny application using the \code{addResourcePath} function from the shiny package. It adds a directory path to the 'temp' folder of the package using the \code{system.file} function.

This function adds a resource path for the shiny application using the \code{addResourcePath} function from the shiny package. It adds a directory path to the 'www' folder of the package using the \code{system.file} function.
}
\examples{
.onLoad(libname = "mylib", pkgname = "mypkg")
.onLoad(libname = "mylib", pkgname = "mypkg")
}
