% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R, R/www.R
\name{.onUnload}
\alias{.onUnload}
\title{Unload Function for interpretablePA Package}
\usage{
.onUnload(libname, pkgname)

.onUnload(libname, pkgname)
}
\arguments{
\item{libname}{character string specifying the path to the library being unloaded.}

\item{pkgname}{character string specifying the name of the package being unloaded.}
}
\description{
This function is called when the interpretablePA package is unloaded. It removes the resource path "temp" previously added with \code{\link{.onLoad}} using the \code{\link[shiny]{removeResourcePath}} function from the shiny package.

This function is called when the interpretablePA package is unloaded. It removes the resource path "www" previously added with \code{\link{.onLoad}} using the \code{\link[shiny]{removeResourcePath}} function from the shiny package.
}
