% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_delta_cvd}
\alias{find_delta_cvd}
\title{Find Delta CVD}
\usage{
find_delta_cvd(mod, myvals, fix = "ACC_day_mg_pla", delta_y_abs = 3.5)
}
\arguments{
\item{mod}{A model object to be used for prediction}

\item{myvals}{A data.frame or matrix containing the input values for prediction}

\item{fix}{A character vector of column names to be held constant during optimization}

\item{delta_y_abs}{The desired absolute change in CVD endpoint}
}
\value{
The delta x value that results in a delta y value of delta_y_abs for the CVD endpoint.
}
\description{
This function finds the delta value for the Cardiovascular Disease (CVD) endpoint for a given model and input values using the uniroot function.
}
\examples{
find_delta_cvd(my_mod, my_vals, fix = "ACC_day_mg_pla", delta_y_abs = 3.5)

}
