% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_delta_x}
\alias{find_delta_x}
\title{Find Delta X}
\usage{
find_delta_x(mod, myvals, fix = "ACC_day_mg_pla", delta_y_perc = 0.1)
}
\arguments{
\item{mod}{a fitted model object}

\item{myvals}{a data frame of predictor variables}

\item{fix}{a character vector specifying the variable to hold constant while finding delta x}

\item{delta_y_perc}{a numeric scalar specifying the percentage increase in the predicted response variable}
}
\value{
a numeric scalar specifying the change in \code{fix} needed to achieve the specified percentage
increase in the predicted response variable
}
\description{
This function finds the change in a specific variable \code{fix} in a data frame \code{myvals}
needed to achieve a specified percentage increase in the predicted response variable
\code{mod}. The function uses the \code{uniroot} function to find the value of \code{x} that
satisfies the equation \code{predict(mod, newdata = tmp_frame) - mygoal = 0}.
}
\examples{
mod <- lm(mpg ~ wt + hp, data = mtcars)
myvals <- data.frame(wt = 2.8, hp = 110)
find_delta_x(mod, myvals, fix = "wt", delta_y_perc = 0.1)

}
