% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_delta_y}
\alias{find_delta_y}
\title{Find Delta Y}
\usage{
find_delta_y(mod, myvals, fix = "", delta_y_abs = 1)
}
\arguments{
\item{mod}{a fitted model object}

\item{myvals}{a data frame of predictor variables}

\item{fix}{a character vector specifying the variable to hold constant while finding delta y}

\item{delta_y_abs}{a numeric scalar specifying the absolute increase in the predicted response variable}
}
\value{
a numeric scalar specifying the change in the predicted response variable needed to achieve
the specified absolute increase in the predicted response variable
}
\description{
This function finds the change in the predicted response variable in a data frame \code{myvals}
needed to achieve a specified absolute increase in the predicted response variable
\code{mod}. The function uses the \code{uniroot} function to find the value of \code{x} that
satisfies the equation \code{predict(mod, newdata = tmp_frame) - mygoal = 0}.
}
\examples{
mod <- lm(mpg ~ wt + hp, data = mtcars)
myvals <- data.frame(wt = 2.8, hp = 110)
find_delta_y(mod, myvals, fix = "", delta_y_abs = 2)

}
