% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_delta_y_ig}
\alias{find_delta_y_ig}
\title{Find Delta Y for intensity gradient}
\usage{
find_delta_y_ig(mod, myvals, fix = "", delta_y_abs = 1)
}
\arguments{
\item{mod}{A fitted GAM or GAMLSS model with inverse Gaussian distribution, as returned by functions like \code{gam()} or \code{gamlss()}.}

\item{myvals}{A data frame containing the predictor values at which the predicted response is to be evaluated. The column names of \code{myvals} should correspond to the variable names used in the model formula.}

\item{fix}{A character vector specifying the names of the variables in \code{myvals} that should be fixed at their original values when searching for the change in \code{fix} that leads to the target change in the predicted response. The default value is an empty character vector, indicating that no variables should be fixed.}

\item{delta_y_abs}{A numeric value specifying the desired absolute change in the predicted response of the model. The default value is 1.}
}
\value{
A numeric value representing the change in the variable \code{fix} that leads to the target change in the predicted response of the model.
}
\description{
Given a fitted GAM or GAMLSS model with inverse Gaussian distribution, and a set of predictor values \code{myvals}, this function finds the change in a specified variable \code{fix} that leads to a target change in the predicted response of the model. The target change is specified by the argument \code{delta_y_abs}, which represents the desired absolute change in the predicted response. If the model is a GAMLSS model, the predicted response is the mean of the response distribution, which is obtained using the \code{what = "mu"} argument of the \code{predict()} function.
}
\examples{
# Fit a GAMLSS model with inverse Gaussian distribution
data(sunspot.year)
mod <- gamlss(log(y) ~ pb(year, 4), data = sunspot.year, family = "IG")

# Find the change in the year variable that leads to a 10\% increase in the predicted response
myvals <- data.frame(year = 1940)
find_delta_y_ig(mod, myvals, fix = "year", delta_y_abs = 0.1*exp(predict(mod, newdata = myvals, what = "mu")))

}
