% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{interpret.pa}
\alias{interpret.pa}
\title{Function Name: interpret.pa()}
\usage{
interpret.pa(...)
}
\arguments{
\item{None}{Not necessary}
}
\description{
The interpret.pa() function is a part of an R application that supports the use of cut-point-free metrics, namely Average Acceleration (AvAcc) and Intensity Gradient (IG), assessed by wrist-worn triaxial accelerometers and calculated over the 24-hour day. This function turns abstract accelerometer data into understandable information by classifying the levels of physical activity based on age- and sex-specific reference values and translating the cut-point-free accelerometer metrics into meaningful outcomes.
}
\examples{
interpret.pa()
The interpret.pa() function returns a Shiny application.
}
\seealso{
\link[vignette:Intro_interpretablePA]{Intro to interpretablePA}
}
\keyword{accelerometer}
\keyword{human_movement}
\keyword{physical_activity}
